// packages/telegram/src/domain/models/group.ts
import type { TelegramPayload } from './payload.js';

export type BasicGroup = {
  botCommands?: TelegramPayload | undefined;
  canHideMembers?: boolean | null | undefined;
  canToggleAggressiveAntiSpam?: boolean | null | undefined;
  creatorUserId?: string | null | undefined;
  description?: string | null | undefined;
  id: string;
  inviteLink?: TelegramPayload | undefined;
  isActive?: boolean | null | undefined;
  memberCount?: number | null | undefined;
  members?: TelegramPayload | undefined;
  photoId?: string | null | undefined;
  status?: TelegramPayload | undefined;
  upgradedToSupergroupId?: string | null | undefined;
};

export type BasicGroupPatch = Partial<BasicGroup> & Pick<BasicGroup, 'id'>;

export type Supergroup = {
  activeStoryState?: TelegramPayload | undefined;
  administratorCount?: number | null | undefined;
  bannedCount?: number | null | undefined;
  boostLevel: number;
  botCommands?: TelegramPayload | undefined;
  botVerification?: TelegramPayload | undefined;
  canEnablePaidMessages?: boolean | null | undefined;
  canEnablePaidReaction?: boolean | null | undefined;
  canGetMembers?: boolean | null | undefined;
  canGetRevenueStatistics?: boolean | null | undefined;
  canGetStarRevenueStatistics?: boolean | null | undefined;
  canGetStatistics?: boolean | null | undefined;
  canHaveSponsoredMessages?: boolean | null | undefined;
  canHideMembers?: boolean | null | undefined;
  canSendGift?: boolean | null | undefined;
  canSetLocation?: boolean | null | undefined;
  canSetStickerSet?: boolean | null | undefined;
  canToggleAggressiveAntiSpam?: boolean | null | undefined;
  customEmojiStickerSetId?: string | null | undefined;
  date: Date;
  description?: string | null | undefined;
  directMessagesChatId?: string | null | undefined;
  giftCount?: number | null | undefined;
  hasAggressiveAntiSpamEnabled?: boolean | null | undefined;
  hasAutomaticTranslation: boolean;
  hasDirectMessagesGroup: boolean;
  hasForumTabs: boolean;
  hasHiddenMembers?: boolean | null | undefined;
  hasLinkedChat: boolean;
  hasLocation: boolean;
  hasPaidMediaAllowed?: boolean | null | undefined;
  hasPinnedStories?: boolean | null | undefined;
  id: string;
  inviteLink?: TelegramPayload | undefined;
  isAdministeredDirectMessagesGroup: boolean;
  isAllHistoryAvailable?: boolean | null | undefined;
  isBroadcastGroup: boolean;
  isChannel: boolean;
  isDirectMessagesGroup: boolean;
  isForum: boolean;
  isSlowModeEnabled: boolean;
  joinByRequest: boolean;
  joinToSendMessages: boolean;
  linkedChatId?: string | null | undefined;
  location?: TelegramPayload | undefined;
  mainProfileTab?: TelegramPayload | undefined;
  memberCount?: number | null | undefined;
  myBoostCount?: number | null | undefined;
  outgoingPaidMessageStarCount?: string | null | undefined;
  paidMessageStarCount: string;
  photoId?: string | null | undefined;
  restrictedCount?: number | null | undefined;
  restrictionInfo?: TelegramPayload | undefined;
  showMessageSender: boolean;
  signMessages: boolean;
  slowModeDelay?: number | null | undefined;
  slowModeDelayExpiresIn?: number | null | undefined;
  status: TelegramPayload;
  stickerSetId?: string | null | undefined;
  unrestrictBoostCount?: number | null | undefined;
  upgradedFromBasicGroupId?: string | null | undefined;
  upgradedFromMaxMessageId?: string | null | undefined;
  usernames?: TelegramPayload | undefined;
  verificationStatus?: TelegramPayload | undefined;
};

export type SupergroupPatch = Partial<Supergroup> & Pick<Supergroup, 'id'>;